+++ App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavigationContainer } from "@react-navigation/native";
-import MainDrawer from './components/Navigator';
-import store from "./redux/store";
+import MainDrawer from './src/navigators/Navigator';
+import store from './src/redux/store';
 import { Provider } from "react-redux";
 
 export default function App() {
