+++ package.json
@@ -49,7 +49,8 @@
     "yarn": "^1.22.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0"
+    "@babel/core": "^7.9.0",
+    "typescript": "~4.0.0"
   },
   "private": true
 }
