+++ src/pages/SignUp/SignUp.jsx
@@ -1,38 +1,65 @@
 import React from "react";
-import { StyleSheet, View, TouchableWithoutFeedback, Keyboard } from "react-native";
+import {
+  StyleSheet,
+  View,
+  TouchableWithoutFeedback,
+  Keyboard,
+  ScrollView,
+} from "react-native";
 import { Text, Button, Input, Switch, Icon } from "react-native-elements";
 import { moderateScale } from "react-native-size-matters";
 //Animations library lottie
 import LottieView from "lottie-react-native";
 //importing animation asset
-import animation from "../../assets/loginAsset.json";
+import animation from "../../../assets/loginAsset.json";
 
-function LoginComponent({
+function SignUpComponent({
   email,
   setEmail,
   setPassword,
   password,
-  switchValue,
-  setSwitchValue,
   handleSubmit,
   emailRegex,
   pwdRegex,
+  name,
+  setName,
+  nameRegex,
+  confirmPassword,
+  setConfirmPassword,
   navigation,
-  handleSwitchChange,
 }) {
   return (
-    <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
+    <ScrollView>
       <View style={styles.container}>
         <View style={styles.titleView}>
           <Text style={styles.title} h2>
-            Sign In
+            Sign Up
           </Text>
-          <Text style={styles.subTitle}>Welcome Back!</Text>
+          <Text style={styles.subTitle}>Excited to have you join us!</Text>
           <View style={styles.animationView}>
             <LottieView autoPlay loop source={animation} />
           </View>
         </View>
         <View style={styles.inputView}>
+          <Input
+            textContentType="username"
+            label="Username"
+            placeholder="e.g johnnyDoe"
+            leftIcon={<Icon name="person" size={21} color="black" />}
+            value={name}
+            onChangeText={setName}
+            errorStyle={{ color: "red" }}
+            errorMessage={
+              //confusing ternary chaining usage so note to self:
+              //checking to see if user has entered any input by looking at the value of the controlling variable
+              name !== ""
+                ? nameRegex.test(name)
+                  ? null
+                  : "Please enter a valid username!"
+                : null
+            
+            }
+          />
           <Input
             textContentType="emailAddress"
             label="Email"
@@ -70,42 +97,47 @@ function LoginComponent({
                 : null
             }
           />
-          <View style={styles.switchView}>
-            <Switch
-              style={styles.switch}
-              value={switchValue}
-              onValueChange={handleSwitchChange}
-            />
-            <Text style={styles.rememberMe}>Remember me</Text>
-            <Text
-              onPress={() => console.log("forgot password pressed")}
-              style={styles.forgotPwd}
-            >
-              Forgot Password
-            </Text>
-          </View>
+          <Input
+            label="Confirm Password"
+            placeholder="Confirm Password"
+            textContentType="password"
+            secureTextEntry
+            leftIcon={<Icon name="lock" size={21} color="black" />}
+            onChangeText={setConfirmPassword}
+            value={confirmPassword}
+            errorStyle={{ color: "red" }}
+            errorMessage={
+              //confusing ternary chaining usage so note to self:
+              //Checks forst to see if confirmpassowrd is not empty then it checks to see if password is not empty
+              //then it goes ahead to check if password is regex compliant if it is then it will check to see if confirmpassword equals password
+              //if it doesnt then it will display password must match error
+              confirmPassword !== ""
+                ? password !== ""
+                  ? pwdRegex.test(password)
+                    ? confirmPassword === password
+                      ? null
+                      : "Passwords must match"
+                    : null
+                  : null
+                : null
+            }
+          />
         </View>
         <View style={styles.buttonView}>
           <Button
             containerStyle={styles.signInButton}
-            title="Sign In"
+            title="Sign Up"
             onPress={handleSubmit}
           />
           <View style={styles.signUpOfferView}>
             <Text style={styles.noAcctTxt}>
-              Don't have an account?
-              <Text
-                style={styles.signUpTxt}
-                onPress={() => navigation.navigate(`Sign Up`)}
-              >
-                {" "}
-                Sign Up
-              </Text>
+              Already have an account?
+              <Text style={styles.signUpTxt} onPress={() => navigation.goBack()}> Sign In</Text>
             </Text>
           </View>
         </View>
       </View>
-    </TouchableWithoutFeedback>
+    </ScrollView>
   );
 }
 
@@ -121,16 +153,15 @@ const styles = StyleSheet.create({
     flexDirection: "column",
     justifyContent: "center",
     alignItems: "center",
+    marginVertical: 25,
   },
   inputView: {
-    flex: 3,
-  },
-  switchView: {
-    flexDirection: "row",
-    justifyContent: "flex-start",
+    flex: 4,
+    marginVertical: 20,
   },
   buttonView: {
     flex: 1.5,
+    marginVertical: 30,
   },
   title: {},
   subTitle: {
@@ -176,4 +207,4 @@ const styles = StyleSheet.create({
   animation: {},
 });
 
-export default LoginComponent;
+export default SignUpComponent;
